+++ typescript/packages/subsurface-viewer/src/layers/triangle/triangleLayer.ts
@@ -1,12 +1,19 @@
+import type React from "react";
+import { isEqual } from "lodash";
+
 import type { UpdateParameters } from "@deck.gl/core/typed";
 import { CompositeLayer } from "@deck.gl/core/typed";
+
+import workerpool from "workerpool";
+
 import type { Material } from "./privateTriangleLayer";
 import PrivateTriangleLayer from "./privateTriangleLayer";
 import type { ExtendedLayerProps } from "../utils/layerTools";
 import type { ReportBoundingBoxAction } from "../../components/Map";
-import { isEqual } from "lodash";
 import { makeFullMesh } from "./webworker";
-import type React from "react";
+
+import config from "../../SubsurfaceConfig.json";
+import { findConfig } from "../../utils/configTools";
 
 export type Params = {
     vertexArray: Float32Array;
@@ -14,6 +21,28 @@ export type Params = {
     smoothShading: boolean;
     displayNormals: boolean;
 };
+
+// init workerpool
+const workerPoolConfig = findConfig(
+    config,
+    "config/workerpool",
+    "config/layer/TriangleLayer/workerpool"
+);
+
+const pool = workerpool.pool({
+    ...{
+        maxWorkers: 10,
+        workerType: "web",
+    },
+    ...workerPoolConfig,
+});
+
+function onTerminateWorker() {
+    const stats = pool.stats();
+    if (stats.busyWorkers === 0 && stats.pendingTasks === 0) {
+        pool.terminate();
+    }
+}
 
 async function loadData(
     pointsData: string | number[] | Float32Array,
@@ -44,7 +73,7 @@ async function loadData(
     }
 
     //-- Triangle indexes --
-    let indexArray: Uint32Array = new Uint32Array();
+    let indexArray: Uint32Array;
     if (Array.isArray(triangleData)) {
         // Input data is native javascript array.
         indexArray = new Uint32Array(triangleData);
@@ -164,12 +193,6 @@ export default class TriangleLayer extends CompositeLayer<TriangleLayerProps> {
         p.then(([vertexArray, indexArray]) => {
             // Using inline web worker for calculating the triangle mesh from
             // loaded input data so not to halt the GUI thread.
-            const blob = new Blob(
-                ["self.onmessage = ", makeFullMesh.toString()],
-                { type: "text/javascript" }
-            );
-            const url = URL.createObjectURL(blob);
-            const webWorker = new Worker(url);
 
             const webworkerParams: Params = {
                 vertexArray,
@@ -178,9 +201,8 @@ export default class TriangleLayer extends CompositeLayer<TriangleLayerProps> {
                 displayNormals: this.props.debug,
             };
 
-            webWorker.postMessage(webworkerParams);
-            webWorker.onmessage = (e) => {
-                const [geometryTriangles, geometryLines] = e.data;
+            pool.exec(makeFullMesh, [{ data: webworkerParams }]).then((e) => {
+                const [geometryTriangles, geometryLines] = e;
 
                 this.setState({
                     geometryTriangles,
@@ -222,13 +244,13 @@ export default class TriangleLayer extends CompositeLayer<TriangleLayerProps> {
                     });
                 }
 
-                webWorker.terminate();
-
                 this.setState({
                     ...this.state,
                     isFinishedLoading: true,
                 });
-            };
+
+                onTerminateWorker();
+            });
         });
     }