+++ typescript/packages/subsurface-viewer/src/storybook/layers/PointsLayer.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { create, all } from "mathjs";
+import { all, create } from "mathjs";
 
 import SubsurfaceViewer from "../../SubsurfaceViewer";
 
@@ -9,6 +9,14 @@ import { default3DViews, defaultStoryParameters } from "../sharedSettings";
 const stories: Meta = {
     component: SubsurfaceViewer,
     title: "SubsurfaceViewer / Points Layer",
+    args: {
+        // Add a reset button for all the stories.
+        // Somehow, I do not manage to add the triggerHome to the general "unset" controls :/
+        triggerHome: 0,
+    },
+
+    // Disable automatic testing of stories that use this tag.
+    tags: ["no-test"],
 };
 export default stories;
 
@@ -159,4 +167,5 @@ export const HugePointsLayer: StoryObj<typeof SubsurfaceViewer> = {
             },
         },
     },
+    tags: ["no-test"],
 };