+++ typescript/packages/subsurface-viewer/src/storybook/layers/TriangleLayer.stories.tsx
@@ -1,4 +1,7 @@
+import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
+
+import { create, all } from "mathjs";
 
 import SubsurfaceViewer from "../../SubsurfaceViewer";
 
@@ -14,6 +17,11 @@ import {
 const stories: Meta = {
     component: SubsurfaceViewer,
     title: "SubsurfaceViewer / Triangle Layer",
+    args: {
+        // Add a reset button for all the stories.
+        // Somehow, I do not manage to add the triggerHome to the general "unset" controls :/
+        triggerHome: 0,
+    },
 };
 export default stories;
 
@@ -221,4 +229,80 @@ export const TypedArrayInput: StoryObj<typeof SubsurfaceViewer> = {
             },
         },
     },
+};
+
+const math = create(all, { randomSeed: "12345" });
+
+const bboxSize = 1000;
+const trglSize = 100;
+
+const randomFunc = (size: number): number => {
+    if (math.random) {
+        return math.random() * size;
+    }
+    return Math.random() * size;
+};
+
+const buildTrgl = (count: number = 1): number[] => {
+    count = count || 1;
+    // 9 is 3 points for the triangle * 3 vertices
+    const trglDataSize = 9;
+    const triangles = Array(trglDataSize * count).fill(0);
+    for (let i = 0; i < count; ++i) {
+        // random triangle center
+        const center = Array(3)
+            .fill(0)
+            .map(() => randomFunc(bboxSize));
+        for (let ti = 0; ti < trglDataSize; ++ti) {
+            triangles[i * trglDataSize + ti] =
+                center[ti % 3] + randomFunc(trglSize);
+        }
+    }
+    return triangles;
+};
+
+const TriangleLayersGenerator: React.FC<{
+    triggerHome: number;
+    layerCount: number;
+    triangleCount: number;
+}> = (props) => {
+    const tsurfLayers = React.useMemo(() => {
+        const result: Record<string, unknown>[] = [];
+        for (let i = 0; i <= props.layerCount; ++i) {
+            result.push({
+                "@@type": "TriangleLayer",
+                id: `triangle-layer-${i}`,
+
+                pointsData: buildTrgl(props.triangleCount),
+
+                triangleData: Array(3 * props.triangleCount)
+                    .fill(0)
+                    .map((_, i) => i),
+
+                //color: [randomFunc(255), randomFunc(255), randomFunc(255)], // Surface color.
+                gridLines: true, // If true will draw lines around triangles.
+                material: true, // If true will use triangle normals for shading.
+                ZIncreasingDownwards: true,
+                //contours: [0, 1],          // If used will display contour lines.
+            });
+        }
+        return result;
+    }, [props.layerCount, props.triangleCount]);
+
+    return (
+        <SubsurfaceViewer
+            triggerHome={props.triggerHome}
+            id="many-triangle-layers"
+            layers={tsurfLayers}
+            views={default3DViews}
+        />
+    );
+};
+
+export const TriangleLayers: StoryObj<typeof TriangleLayersGenerator> = {
+    args: {
+        layerCount: 10,
+        triangleCount: 1000,
+    },
+    render: (args) => <TriangleLayersGenerator {...args} />,
 };