+++ typescript/packages/subsurface-viewer/src/storybook/layers/WellsLayer.stories.tsx
@@ -1,25 +1,25 @@
 /* eslint-disable react-hooks/exhaustive-deps */ // remove when ready to fix these.
 
+import type { Meta, StoryObj } from "@storybook/react";
 import type { SyntheticEvent } from "react";
 import React, { useState } from "react";
-import type { Meta, StoryObj } from "@storybook/react";
 
-import type { FeatureCollection } from "geojson";
 import { Slider } from "@mui/material";
 import { styled } from "@mui/material/styles";
+import type { FeatureCollection } from "geojson";
 
-import { NativeSelect } from "@equinor/eds-core-react";
 import {
-    createColorMapFunction,
     ColorLegend,
     colorTables,
+    createColorMapFunction,
 } from "@emerson-eps/color-tables";
+import { NativeSelect } from "@equinor/eds-core-react";
 
 import type { SubsurfaceViewerProps } from "../../SubsurfaceViewer";
 import SubsurfaceViewer from "../../SubsurfaceViewer";
 import type { MapMouseEvent } from "../../components/Map";
-import WellsLayer from "../../layers/wells/wellsLayer";
 import AxesLayer from "../../layers/axes/axesLayer";
+import WellsLayer from "../../layers/wells/wellsLayer";
 
 import {
     default3DViews,
@@ -32,6 +32,12 @@ import {
 const stories: Meta = {
     component: SubsurfaceViewer,
     title: "SubsurfaceViewer / Wells Layer",
+    args: {
+        // Add a reset button for all the stories.
+        // Somehow, I do not manage to add the triggerHome to the general "unset" controls :/
+        triggerHome: 0,
+    },
+    tags: ["no-test"],
 };
 export default stories;
 
@@ -47,15 +53,16 @@ const testWellWithDuplicates = {
                 geometries: [
                     {
                         type: "Point",
-                        coordinates: [0, 0, 0],
+                        coordinates: [0, 0],
                     },
                     {
                         type: "LineString",
                         coordinates: [
-                            [0, 0, 0],
                             [0, 0, -100],
                             [0, 0, -200],
                             [0, 0, -300],
+                            [0, 0, -400],
+                            [0, 0, -400],
                             [0, 0, -400],
                             [0, 0, -500],
                             [0, 0, -600],
@@ -159,6 +166,7 @@ export const volveWells2: StoryObj<typeof SubsurfaceViewer> = {
         ],
     },
     render: (args) => <VolveWellsWithMouseCallback {...args} />,
+    tags: ["no-test"],
 };
 
 // Volve wells with logs.
@@ -546,6 +554,7 @@ export const WellsRefine: StoryObj<typeof WellsRefineComponent> = {
         },
     },
     render: (args) => <WellsRefineComponent {...args} />,
+    tags: ["no-test"],
 };
 
 export const Wells3d: StoryObj<typeof SubsurfaceViewer> = {