+++ typescript/packages/wsc-common/CHANGELOG.md
@@ -1,3 +1,31 @@
+## [0.5.3](https://github.com/equinor/webviz-subsurface-components/compare/wsc-common@0.5.2...wsc-common@0.5.3) (2024-02-16)
+
+
+### Bug Fixes
+
+* bump [@deck](https://github.com/deck).gl/core from 8.9.33 to 8.9.34 in /typescript ([#1922](https://github.com/equinor/webviz-subsurface-components/issues/1922)) ([94bd152](https://github.com/equinor/webviz-subsurface-components/commit/94bd1524b7cba4d505e559ef9648981b7fdebbca)), closes [#8412](https://github.com/equinor/webviz-subsurface-components/issues/8412) [#8290](https://github.com/equinor/webviz-subsurface-components/issues/8290) [#8347](https://github.com/equinor/webviz-subsurface-components/issues/8347) [#8412](https://github.com/equinor/webviz-subsurface-components/issues/8412) [#8290](https://github.com/equinor/webviz-subsurface-components/issues/8290) [#8347](https://github.com/equinor/webviz-subsurface-components/issues/8347) [#8436](https://github.com/equinor/webviz-subsurface-components/issues/8436) [#8412](https://github.com/equinor/webviz-subsurface-components/issues/8412) [#8290](https://github.com/equinor/webviz-subsurface-components/issues/8290)
+
+## [0.5.2](https://github.com/equinor/webviz-subsurface-components/compare/wsc-common@0.5.1...wsc-common@0.5.2) (2024-02-13)
+
+
+### Bug Fixes
+
+* bump @equinor/eds-icons from 0.19.3 to 0.21.0 and @equinor/eds-core-react from 0.33.0 to 0.36.0 in /typescript ([#1892](https://github.com/equinor/webviz-subsurface-components/issues/1892)) ([c7bd611](https://github.com/equinor/webviz-subsurface-components/commit/c7bd611f09958df9f8c58006bd3fb1a675f2ca49)), closes [equinor/design-system#2367](https://github.com/equinor/design-system/issues/2367) [equinor/design-system#2431](https://github.com/equinor/design-system/issues/2431) [equinor/design-system#2378](https://github.com/equinor/design-system/issues/2378) [equinor/design-system#2399](https://github.com/equinor/design-system/issues/2399) [equinor/design-system#2410](https://github.com/equinor/design-system/issues/2410) [equinor/design-system#2432](https://github.com/equinor/design-system/issues/2432) [equinor/design-system#2442](https://github.com/equinor/design-system/issues/2442) [equinor/design-system#2420](https://github.com/equinor/design-system/issues/2420) [equinor/design-system#2377](https://github.com/equinor/design-system/issues/2377) [equinor/design-system#2384](https://github.com/equinor/design-system/issues/2384) [equinor/design-system#2405](https://github.com/equinor/design-system/issues/2405) [equinor/design-system#2460](https://github.com/equinor/design-system/issues/2460) [equinor/design-system#2247](https://github.com/equinor/design-system/issues/2247) [equinor/design-system#2436](https://github.com/equinor/design-system/issues/2436) [equinor/design-system#2451](https://github.com/equinor/design-system/issues/2451) [equinor/design-system#2303](https://github.com/equinor/design-system/issues/2303) [equinor/design-system#2327](https://github.com/equinor/design-system/issues/2327) [equinor/design-system#2337](https://github.com/equinor/design-system/issues/2337) [equinor/design-system#2335](https://github.com/equinor/design-system/issues/2335) [equinor/design-system#2313](https://github.com/equinor/design-system/issues/2313) [equinor/design-system#3177](https://github.com/equinor/design-system/issues/3177) [#3239](https://github.com/equinor/webviz-subsurface-components/issues/3239) [#3219](https://github.com/equinor/webviz-subsurface-components/issues/3219) [#3177](https://github.com/equinor/webviz-subsurface-components/issues/3177) [#3137](https://github.com/equinor/webviz-subsurface-components/issues/3137) [#3132](https://github.com/equinor/webviz-subsurface-components/issues/3132) [#3121](https://github.com/equinor/webviz-subsurface-components/issues/3121) [#3020](https://github.com/equinor/webviz-subsurface-components/issues/3020) [#3019](https://github.com/equinor/webviz-subsurface-components/issues/3019)
+
+## [0.5.1](https://github.com/equinor/webviz-subsurface-components/compare/wsc-common@0.5.0...wsc-common@0.5.1) (2024-02-08)
+
+
+### Bug Fixes
+
+* bump [@deck](https://github.com/deck).gl/react from 8.9.33 to 8.9.34 in /typescript ([#1911](https://github.com/equinor/webviz-subsurface-components/issues/1911)) ([ac4e20c](https://github.com/equinor/webviz-subsurface-components/commit/ac4e20cd1ae321612ad4aff52f90fd5c6de3af97)), closes [#8412](https://github.com/equinor/webviz-subsurface-components/issues/8412) [#8290](https://github.com/equinor/webviz-subsurface-components/issues/8290) [#8347](https://github.com/equinor/webviz-subsurface-components/issues/8347) [#8412](https://github.com/equinor/webviz-subsurface-components/issues/8412) [#8290](https://github.com/equinor/webviz-subsurface-components/issues/8290) [#8347](https://github.com/equinor/webviz-subsurface-components/issues/8347) [#8436](https://github.com/equinor/webviz-subsurface-components/issues/8436) [#8412](https://github.com/equinor/webviz-subsurface-components/issues/8412) [#8290](https://github.com/equinor/webviz-subsurface-components/issues/8290)
+
+# [0.5.0](https://github.com/equinor/webviz-subsurface-components/compare/wsc-common@0.4.4...wsc-common@0.5.0) (2024-02-05)
+
+
+### Features
+
+* use webworker pool to control number of workers ([#1905](https://github.com/equinor/webviz-subsurface-components/issues/1905)) ([304c920](https://github.com/equinor/webviz-subsurface-components/commit/304c920fa229bd8bdd70102e48b6a2445c0cd234))
+
 ## [0.4.4](https://github.com/equinor/webviz-subsurface-components/compare/wsc-common@0.4.3...wsc-common@0.4.4) (2024-01-19)